+++ brewtils/auto_decorator.py
@@ -1,4 +1,6 @@
-import inspect
+from inspect import Parameter as InspectParameter  # noqa
+from inspect import signature
+
 
 from brewtils.models import Command, Parameter
 
@@ -27,8 +29,11 @@ class AutoDecorator:
                 _wrapped = getattr(client, func)
                 if not hasattr(_wrapped, "_command") and not func.startswith("__"):
                     # decorators.py will handle all of the markings
+                    has_kwargs = any([True for p in signature(_wrapped).parameters.values() if p.kind == InspectParameter.VAR_KEYWORD])
                     if func.startswith("_"):
                         _wrapped._command = Command(hidden=True)
+                    elif has_kwargs:
+                        _wrapped._command = Command(allow_any_kwargs=True)
                     else:
                         _wrapped._command = Command()