+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Brewtils Changelog
 TBD
 
 - Fixed cross-server url prefix comparison and handled case where there is no current request
+- Added API support for Latest System, SystemClient will use Version `latest` instead of resolved version. 
+  Allowing Beer Garden to resolve the latest version.
+- Must upgrade to a minimum version of Beer Garden 3.26.0 to support new APIs
 
 3.25.1
 ------