+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Brewtils Changelog
 ------
 TBD
 
+- Added support for autobrew any kwargs
 - Added API support for Latest System, SystemClient will use Version `latest` instead of resolved version. 
   Allowing Beer Garden to resolve the latest version.
 - Must upgrade to a minimum version of Beer Garden 3.26.0 to support new APIs