+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Brewtils Changelog
 TBD
 
 - Added support for autobrew any kwargs
+- Fixed cross-server url prefix comparison and handled case where there is no current request
 - Added API support for Latest System, SystemClient will use Version `latest` instead of resolved version. 
   Allowing Beer Garden to resolve the latest version.
 - Must upgrade to a minimum version of Beer Garden 3.26.0 to support new APIs